Add support for DMD 2.094 -preview=in switch
This will allow users to use -preview=in as soon as it's released.
'in ref' on string doesn't make much sense, so it was turned into a simple 'in'.
'Location' is a string + an 'int', and while that would be passed by ref with the
new switch, there's not much point in ensuring ref here, as the difference is
minimal, but prevents using rvalues, so it was also turned into a simple 'in'.
Lastly, for node, it was changed to be simply 'const ref', so as to keep the 'ref'.